feat: press-header.js 파일 분리
