+++ src/index.js
@@ -1,11 +1,12 @@
 import { setNewsStandHeader } from './components/news-stand-header.js';
 import { setLatestNews } from './components/latest-news.js';
-import { setPressGrid } from './components/press-grid.js';
+import { setPressHeader } from './components/press-header.js';
 
 let isLightMode = true;
+let isGridMode = true;
 
 (function init() {
   setNewsStandHeader();
   setLatestNews();
-  setPressGrid(isLightMode);
+  setPressHeader(isLightMode);
 })();
